+++ Tortuga.Chain/Changelog.md
@@ -8,6 +8,7 @@
 * [#371 Implement GetForeignKeysForTable for PostgreSQL](https://github.com/TortugaResearch/Tortuga.Chain/issues/371)
 * [#372 Implement GetForeignKeysForTable for MySQL](https://github.com/TortugaResearch/Tortuga.Chain/issues/372)
 * [#374 Implement GetForeignKeysForTable for SQLite](https://github.com/TortugaResearch/Tortuga.Chain/issues/374)
+* [#373 Implement GetForeignKeysForTable for Access](https://github.com/TortugaResearch/Tortuga.Chain/issues/373)
 
 You can now call `table.GetForeignKeys()` on tables.